fixed the paymelt quote funcitons.
